+++ src/template_renderer.rs
@@ -4,6 +4,7 @@ use anyhow::bail;
 use saphyr::Yaml;
 use tera::{self, Tera, Value};
 
+use crate::builtins::add_builtins;
 use crate::confluence_client::ConfluenceClient;
 use crate::error::Result;
 use crate::frontmatter::FrontMatter;
@@ -13,74 +14,6 @@ use crate::mentions::make_mention;
 
 pub struct TemplateRenderer {
     tera: Tera,
-}
-
-fn hello_world(
-    _args: &HashMap<String, serde_json::Value>,
-) -> std::result::Result<serde_json::Value, tera::Error> {
-    Ok(serde_json::to_value("<em>hello world!</em>").unwrap())
-}
-
-fn toc(
-    _args: &HashMap<String, serde_json::Value>,
-) -> std::result::Result<serde_json::Value, tera::Error> {
-    Ok(
-        serde_json::to_value(
-        r#"<ac:structured-macro ac:name="toc" ac:schema-version="1" data-layout="default" ac:macro-id="334277ff-40b1-45ec-b5c7-ba6091fd0df3">
-        <ac:parameter ac:name="minLevel">1</ac:parameter>
-        <ac:parameter ac:name="maxLevel">6</ac:parameter>
-        <ac:parameter ac:name="include" />
-        <ac:parameter ac:name="outline">false</ac:parameter>
-        <ac:parameter ac:name="indent" />
-        <ac:parameter ac:name="exclude" />
-        <ac:parameter ac:name="type">list</ac:parameter>
-        <ac:parameter ac:name="class" />
-        <ac:parameter ac:name="printable">false</ac:parameter>
-    </ac:structured-macro>"#).unwrap()
-    )
-}
-
-fn children(
-    _args: &HashMap<String, serde_json::Value>,
-) -> std::result::Result<serde_json::Value, tera::Error> {
-    Ok(
-        serde_json::to_value(
-        r#"<ac:structured-macro ac:name="children" ac:schema-version="2" data-layout="default" ac:macro-id="4172775450124db364aa2f7e7faf4cb3" />"#
-        ).unwrap()
-    )
-}
-
-fn labellist(
-    args: &HashMap<String, serde_json::Value>,
-) -> std::result::Result<serde_json::Value, tera::Error> {
-    let label = args
-        .get(&"labels".to_string())
-        .ok_or("Missing required argument 'labels'")?;
-
-    let parameter = match label {
-        serde_json::Value::String(s) => format!("label = \"{}\"", s),
-        serde_json::Value::Array(a) => {
-            if !a.is_empty() {
-                format!(
-                    "label in ({})",
-                    a.iter()
-                        .map(|s| s.to_string())
-                        .collect::<Vec<String>>()
-                        .join(",")
-                )
-            } else {
-                Err(tera::Error::msg("labels needs to be a non-empty array"))?
-            }
-        }
-        _ => Err(tera::Error::msg("labels needs to be a string or array"))?,
-    };
-    Ok(
-        serde_json::to_value(
-format!(r#"<ac:structured-macro ac:name="contentbylabel" ac:schema-version="4" data-layout="default" ac:macro-id="808ece5f-14fd-4c2d-853a-bf87e0696e48">
-    <ac:parameter ac:name="cql">{} and space = currentSpace()</ac:parameter>
-</ac:structured-macro>"#, parameter) 
-        ).unwrap()
-    )
 }
 
 fn make_metadata_lookup(metadata: Yaml) -> impl tera::Function {
@@ -103,28 +36,21 @@ fn make_metadata_lookup(metadata: Yaml) -> impl tera::Function {
     )
 }
 
+// Required method
 impl TemplateRenderer {
     pub fn new(space: &MarkdownSpace, client: &ConfluenceClient) -> Result<TemplateRenderer> {
         let mut tera = Tera::new(space.dir.join("**/*.md").into_os_string().to_str().unwrap())?;
 
-        Self::add_builtins(&mut tera)?;
+        add_builtins(&mut tera);
         tera.register_function("mention", make_mention(client.clone()));
 
         Ok(TemplateRenderer { tera })
     }
 
-    fn add_builtins(tera: &mut Tera) -> Result<()> {
-        tera.register_function("hello_world", hello_world);
-        tera.register_function("toc", toc);
-        tera.register_function("children", children);
-        tera.register_function("labellist", labellist);
-        Ok(())
-    }
-
     #[cfg(test)]
     pub fn default() -> Result<TemplateRenderer> {
         let mut tera = Tera::default();
-        Self::add_builtins(&mut tera)?;
+        add_builtins(&mut tera);
 
         Ok(TemplateRenderer { tera })
     }
@@ -132,7 +58,7 @@ impl TemplateRenderer {
     #[cfg(test)]
     pub fn default_with_client(client: &ConfluenceClient) -> Result<TemplateRenderer> {
         let mut tera = Tera::default();
-        Self::add_builtins(&mut tera)?;
+        add_builtins(&mut tera);
 
         tera.register_function("mention", make_mention(client.clone()));
 
@@ -180,13 +106,11 @@ impl TemplateRenderer {
 
 #[cfg(test)]
 mod test {
-    use std::collections::HashMap;
-
     use saphyr::Yaml;
 
     use crate::{error::TestResult, frontmatter::FrontMatter};
 
-    use super::{labellist, TemplateRenderer};
+    use super::TemplateRenderer;
 
     #[test]
     fn it_puts_original_filename_in_message() -> TestResult {
@@ -201,27 +125,6 @@ mod test {
             format!("{:#}", result.unwrap_err()),
             "Failed to render '__tera_one_off': Function 'func_does_not_exist' not found"
         );
-        Ok(())
-    }
-
-    #[test]
-    fn labellist_allows_multiple_labels() -> TestResult {
-        let args = HashMap::from([("labels".to_string(), serde_json::Value::from("foo"))]);
-        let result = labellist(&args);
-
-        assert!(result.is_ok());
-
-        let args = HashMap::from([(
-            "labels".to_string(),
-            serde_json::Value::from(vec!["foo", "bar"]),
-        )]);
-        let result = labellist(&args);
-
-        assert!(result.is_ok());
-        let s = result.unwrap();
-        println!("s: {:#}", s);
-        assert!(s.to_string().contains(r#"label in (\"foo\",\"bar\")"#));
-
         Ok(())
     }
